requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.2.10           # via django
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
django-crontab==0.7.1
django-filter==2.3.0
django==3.0.8             # via django-crontab, django-filter, djangorestframework
djangorestframework==3.11.0
idna==2.10                # via requests
importlib-metadata==1.7.0  # via markdown
markdown==3.2.2
numpy==1.19.0             # via pandas
pandas==1.0.5             # via youtube-data-api
python-dateutil==2.8.1    # via pandas
pytz==2020.1              # via django, pandas
requests==2.24.0          # via youtube-data-api
six==1.15.0               # via python-dateutil
sqlparse==0.3.1           # via django
urllib3==1.25.9           # via requests
youtube-data-api==0.0.20
zipp==3.1.0               # via importlib-metadata
